RenderPass: Drop early-out for passes without attachments (#19196)
After #18237 we are no longer able to use render passes without output
attachments (i.e. raster passes that only write to sideband buffers)
following an assert inside `GetRenderAttachmentConfiguration()`.  This
was supposedly explicitly supported before and seems to be caused by an
unneeded early-out when an empty attachment configuration was supposed
to be built inside `BuildRenderAttachmentConfiguration()`.

Signed-off-by: Marijn Suijten <marijn@traverseresearch.nl>