Cherry-pic `need-sync:to-development` changes from `stabilization` to `development` (#19277)
* Cherry-pick from `stabilization` - clang 18 compile fix

Fixes a compile issue specific to clang 18 (#19247)

* Fixes a compile issue specific to clang 18

There is an issue with clang 18 where it does not
compile correctly when using conditional explicit.

This bug was fixed above clang 19, but affects clang 18
including versions of the android NDK.

This switches off the use of that feature specifically for
clang 18.

---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Cherry-pick from stabilization - cmake to 3.24.x  (#19268)

* Updates the engine to require CMake >= 3.24

Updates the WIX installer script and build script to include that version
of CMake downloaded from the kitware official site.

Makes the WIX post-installation steps non-failing.  They still run,
but if they fail, they do not cause the installer to rollback.

Updates the license scanner to work when licenses live in multiple
different drives on windows.  relpath in python throws an exception
if you specify paths on different drives.  This is often the case
with the 3rd party libraries (`%USERPROFILE%`) and the build
`D:/workspace` or something.

---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>